moved grid_mapping.py to projects directory.  This is the file before a lot of hopefully unneeded code.
